model run bug: check the empty resource and show the correct error info

+++ tethysapp/ueb_app/controllers.py
@@ -310,52 +310,62 @@ def model_run_submit_execution(request):
 
         # move all files in the same folder
         model_input_folder = os.path.join(temp_dir, res_id,  'data', 'contents')
-        model_files_path_list = [os.path.join(model_input_folder, name) for name in os.listdir(model_input_folder)]
 
-        while model_files_path_list:
-            added_files_list = []
+        if os.path.isdir(model_input_folder):
+            model_files_path_list = [os.path.join(model_input_folder, name) for name in os.listdir(model_input_folder)]
 
-            for model_file_path in model_files_path_list:
+            # unzip .zip and move folder files to the same temp_dir level
+            while model_files_path_list:
+                added_files_list = []
 
-                if os.path.isfile(model_file_path) and os.path.splitext(model_file_path)[1] == '.zip':
-                    zf = zipfile.ZipFile(model_file_path, 'r')
-                    zf.extractall(model_input_folder)
-                    extract_file_names = zf.namelist()
-                    added_files_list += [os.path.join(model_input_folder, name) for name in extract_file_names]
-                    zf.close()
-                    os.remove(model_file_path)
+                for model_file_path in model_files_path_list:
 
-                elif os.path.isdir(model_file_path):
-                    for dirpath, _, filenames in os.walk(model_file_path):
-                        for name in filenames:
-                            sub_file_path = os.path.abspath(os.path.join(dirpath, name))
-                            new_file_path = os.path.join(model_input_folder, name)
-                            shutil.move(sub_file_path, new_file_path)
-                            added_files_list.append(new_file_path)
-                    shutil.rmtree(model_file_path)
+                    if os.path.isfile(model_file_path) and os.path.splitext(model_file_path)[1] == '.zip':
+                        zf = zipfile.ZipFile(model_file_path, 'r')
+                        zf.extractall(model_input_folder)
+                        extract_file_names = zf.namelist()
+                        added_files_list += [os.path.join(model_input_folder, name) for name in extract_file_names]
+                        zf.close()
+                        os.remove(model_file_path)
 
-            model_files_path_list = added_files_list
+                    elif os.path.isdir(model_file_path):
+                        for dirpath, _, filenames in os.walk(model_file_path):
+                            for name in filenames:
+                                sub_file_path = os.path.abspath(os.path.join(dirpath, name))
+                                new_file_path = os.path.join(model_input_folder, name)
+                                shutil.move(sub_file_path, new_file_path)
+                                added_files_list.append(new_file_path)
+                        shutil.rmtree(model_file_path)
+
+                model_files_path_list = added_files_list
 
 
-        # TODO: model parameter and data files check
+            # TODO: model parameter and data files check
 
-        # upload the model files to HydroDS
-        model_files_path_list = [os.path.join(model_input_folder, name) for name in os.listdir(model_input_folder)]
-        zip_file_path = os.path.join(model_input_folder, 'input_package.zip')
-        zf = zipfile.ZipFile(zip_file_path, 'w')
-        for file_path in model_files_path_list:
-            zf.write(file_path)
-        zf.close()
-        upload_zip_file_url = client.upload_file(file_to_upload=zip_file_path)
-        client.delete_my_file(upload_zip_file_url.split('/')[-1])
+            # upload the model input and parameter files to HydroDS
+            model_files_path_list = [os.path.join(model_input_folder, name) for name in os.listdir(model_input_folder)]
+            zip_file_path = os.path.join(model_input_folder, 'input_package.zip')
+            zf = zipfile.ZipFile(zip_file_path, 'w')
+            for file_path in model_files_path_list:
+                zf.write(file_path)
+            zf.close()
+            upload_zip_file_url = client.upload_file(file_to_upload=zip_file_path)
+            client.delete_my_file(upload_zip_file_url.split('/')[-1]) #TODO clean this line for testing
 
-        # TODO: call model run service
+            # TODO: call model run service
 
-        ajax_response = {
-            'status': 'Success',
-            'result': upload_zip_file_url
-        }
+            ajax_response = {
+                'status': 'Success',
+                'result': upload_zip_file_url
+            }
+        else:
+            ajax_response = {
+                'status': 'Error',
+                'result': 'The model instance resource includes no model input data and parameter files.'
+            }
 
+        # remove the tempdir
+        shutil.rmtree(temp_dir)
 
     except Exception as e:
         ajax_response = {
